+++ db/migrations/1722676430521-Views.js
@@ -1,10 +1,17 @@
 
 const { getViewDefinitions } = require('../viewDefinitions')
 
-module.exports = class Views1710699134452 {
-  name = 'Views1710699134452'
+module.exports = class Views1722676430521 {
+  name = 'Views1722676430521'
 
   async up(db) {
+    // these two queries will be invoked and the cleaned up by the squid itself
+    // we only do this to be able to reference processor height in mappings 
+    await db.query(`CREATE SCHEMA IF NOT EXISTS squid_processor;`)
+    await db.query(`CREATE TABLE IF NOT EXISTS squid_processor.status (
+      id SERIAL PRIMARY KEY,
+      height INT
+    );`)
     const viewDefinitions = getViewDefinitions(db);
     for (const [tableName, viewConditions] of Object.entries(viewDefinitions)) {
       if (Array.isArray(viewConditions)) {