+++ src/server-extension/resolvers/CreatorToken/types.ts
@@ -1,4 +1,10 @@
-import { ArgsType, Field, Int, ObjectType } from 'type-graphql'
+import { ArgsType, Field, Float, Int, ObjectType } from 'type-graphql'
+import {
+  CreatorToken,
+  MarketplaceTokenOrderByInput,
+  MarketplaceTokenWhereInput,
+  TokenWhereInput,
+} from '../baseTypes'
 
 @ArgsType()
 export class GetShareDividensArgs {
@@ -43,4 +49,80 @@ export class GetAccountTransferrableBalanceArgs {
 export class GetAccountTransferrableBalanceResult {
   @Field(() => Int, { nullable: false })
   transferrableCrtAmount!: number
+}
+
+@ObjectType()
+export class MarketplaceTokensReturnType {
+  @Field(() => CreatorToken, { nullable: false }) creatorToken!: CreatorToken
+  @Field(() => Float, { nullable: false }) pricePercentageChange!: number
+}
+
+@ObjectType()
+export class TopSellingTokensReturnType {
+  @Field(() => CreatorToken, { nullable: false }) creatorToken!: CreatorToken
+  @Field(() => String, { nullable: false }) ammVolume!: string
+}
+
+@ArgsType()
+export class MarketplaceTokensArgs {
+  @Field(() => TokenWhereInput, { nullable: true })
+  where?: Record<string, unknown>
+
+  @Field(() => Int, {
+    nullable: false,
+    description: 'The number of days in period',
+  })
+  periodDays: number
+
+  @Field(() => Int, {
+    nullable: true,
+  })
+  limit?: number
+
+  @Field(() => Boolean, {
+    nullable: true,
+    description: 'Whether the result should be order by price change descending',
+  })
+  orderByPriceDesc: boolean | null
+}
+
+@ObjectType()
+export class MarketplaceTokenCount {
+  @Field(() => Int, { nullable: false }) count: number
+}
+
+@ArgsType()
+export class MarketplaceTokensCountArgs {
+  @Field(() => MarketplaceTokenWhereInput, { nullable: true })
+  where?: Record<string, unknown>
+}
+
+@ArgsType()
+export class MarketplaceTableTokensArgs {
+  @Field(() => MarketplaceTokenWhereInput, { nullable: true })
+  where?: Record<string, unknown>
+
+  @Field(() => Int, {
+    nullable: true,
+    defaultValue: 10,
+    description: 'The number of videos to return',
+  })
+  limit?: number
+
+  @Field(() => Int, {
+    nullable: true,
+  })
+  offset?: number
+
+  @Field(() => [MarketplaceTokenOrderByInput], {
+    nullable: true,
+    description: 'Order of input',
+  })
+  orderBy?: unknown[]
+}
+
+@ArgsType()
+export class TopSellingTokensArgs {
+  @Field(() => TokenWhereInput, { nullable: true })
+  where?: Record<string, unknown>
 }