+++ src/tests/integration/notifications.test.ts
@@ -38,7 +38,7 @@ import {
   OFFCHAIN_NOTIFICATION_ID_TAG,
   RUNTIME_NOTIFICATION_ID_TAG,
 } from '../../utils/notification/helpers'
-import { EntityManagerOverlay } from '../../utils/overlay'
+import { AnyEntity, Constructor, EntityManagerOverlay } from '../../utils/overlay'
 import { defaultTestBlock, populateDbWithSeedData } from './testUtils'
 
 dontenvConfig({
@@ -49,9 +49,19 @@ const metadataToBytes = <T>(metaClass: AnyMetadataClass<T>, obj: T): Uint8Array 
   return Buffer.from(metaClass.encode(obj).finish())
 }
 
-const getNextNotificationId = async (em: EntityManager, onchain: boolean) => {
+const getNextNotificationId = async (
+  store: EntityManager | EntityManagerOverlay,
+  onchain: boolean
+) => {
   const tag = onchain ? RUNTIME_NOTIFICATION_ID_TAG : OFFCHAIN_NOTIFICATION_ID_TAG
-  const row = await em.getRepository(NextEntityId).findOneBy({ entityName: tag })
+  if (store instanceof EntityManagerOverlay) {
+    const row = await store
+      .getRepository(NextEntityId as Constructor<NextEntityId & AnyEntity>)
+      .getOneBy({ entityName: tag })
+    const id = parseInt(row?.nextId.toString() || '1')
+    return id
+  }
+  const row = await store.getRepository(NextEntityId).findOneBy({ entityName: tag })
   const id = parseInt(row?.nextId.toString() || '1')
   return id
 }
@@ -68,6 +78,7 @@ describe('notifications tests', () => {
   let em: EntityManager
   before(async () => {
     em = await globalEm
+    overlay = await createOverlay()
     await populateDbWithSeedData()
   })
   describe('👉 YPP Verify channel', () => {
@@ -272,10 +283,9 @@ describe('notifications tests', () => {
 
     before(async () => {
       const bidAmount = BigInt(100000)
-      nextNotificationIdPre = await getNextNotificationId(em, true)
+      nextNotificationIdPre = await getNextNotificationId(overlay, true)
       notificationId = RUNTIME_NOTIFICATION_ID_TAG + '-' + nextNotificationIdPre
       nft = await em.getRepository(OwnedNft).findOneByOrFail({ videoId })
-      overlay = await createOverlay()
 
       await auctionBidMadeInner(
         overlay,
@@ -355,8 +365,7 @@ describe('notifications tests', () => {
       asV2001: ['3', metadataToBytes(MemberRemarked, metadataMessage), undefined],
     } as any
     before(async () => {
-      overlay = await createOverlay()
-      nextNotificationIdPre = await getNextNotificationId(em, true)
+      nextNotificationIdPre = await getNextNotificationId(overlay, true)
       notificationId = RUNTIME_NOTIFICATION_ID_TAG + '-' + nextNotificationIdPre.toString()
     })
     it('should process video liked and deposit notification', async () => {
@@ -368,7 +377,7 @@ describe('notifications tests', () => {
         event,
       })
 
-      const nextNotificationId = await getNextNotificationId(em, true)
+      const nextNotificationId = await getNextNotificationId(overlay, true)
       notification = (await overlay
         .getRepository(Notification)
         .getByIdOrFail(notificationId)) as Notification
@@ -409,8 +418,7 @@ describe('notifications tests', () => {
       asV2001: ['2', metadataToBytes(MemberRemarked, metadataMessage), undefined], // avoid comment author == creator
     } as any
     before(async () => {
-      overlay = await createOverlay()
-      nextNotificationIdPre = await getNextNotificationId(em, true)
+      nextNotificationIdPre = await getNextNotificationId(overlay, true)
       notificationId = RUNTIME_NOTIFICATION_ID_TAG + '-' + nextNotificationIdPre.toString()
     })
     it('should process comment to video and deposit notification', async () => {
@@ -422,7 +430,7 @@ describe('notifications tests', () => {
         event,
       })
 
-      const nextNotificationId = await getNextNotificationId(em, true)
+      const nextNotificationId = await getNextNotificationId(overlay, true)
       notification = (await overlay
         .getRepository(Notification)
         .getByIdOrFail(notificationId)) as Notification | null
@@ -472,7 +480,7 @@ describe('notifications tests', () => {
       } as any
 
       before(async () => {
-        nextNotificationIdPre = await getNextNotificationId(em, true)
+        nextNotificationIdPre = await getNextNotificationId(overlay, true)
         notificationId = RUNTIME_NOTIFICATION_ID_TAG + '-' + nextNotificationIdPre.toString()
 
         await processMemberRemarkedEvent({
@@ -487,7 +495,7 @@ describe('notifications tests', () => {
       describe('should process reply to comment and deposit notification', () => {
         let nextNotificationId: number
         before(async () => {
-          nextNotificationId = await getNextNotificationId(em, true)
+          nextNotificationId = await getNextNotificationId(overlay, true)
           notification = (await overlay
             .getRepository(Notification)
             .getByIdOrFail(notificationId)) as Notification | null